OneDrive/Desktop/klashibot/requirements.txt
# Core dependencies
kalshi-python>=1.0.0
requests>=2.31.0
websockets>=11.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Async and concurrency
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0

# Database and storage
sqlalchemy>=2.0.0
firebase-admin>=6.2.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# PayPal SDK for automatic transfers
paypalrestsdk>=1.13.3
